Refactoring. Page controller
