+++ controller/ControllerUpload.php
@@ -1,0 +1,7 @@
+<?php
+//todo: include
+function controllerUpload(){
+    $res = model();
+    $v = view($res, 'upload_form.tpl');
+    render($v);
+}
