+++ index.php
@@ -1,47 +1,15 @@
 <?php
-echo "Stage1"; // уровень отладки кода
-function render($data) {
-    echo $data;
-}
+$path = './session';
+session_save_path($path);
+session_start();
+require_once "./model/ModelIndex.php";
+//require_once "./controller/ControllerLogin.php";
+require_once "./controller/ControllerPage.php";
 
-function controller_login(){
-    $res = model();
-    $v = view($res, 'form_login.tpl');
-    render($v);
-}
-
-function controller_upload(){
-    $res = model();
-    $v = view($res, 'upload_form.tpl');
-    render($v);
-}
-// Контроллер
-function controller_index(){
-    $res = model();
-    $v = view($res, 'upload_form.tpl');
-    render($v);
-}
-// Mодель
-function model() {
-      // 1. Получить данные от сервера БД, файла,
-      //http , сайта, API
-    $arr = ['title'=> 'О нас','name'=>"Alisa"];
-    return $arr;
-}
-echo "Stage2";  // уровень отладки кода
-// Представление
-function view($data, $tpl){
-    extract($data);
-    $title = '<h1>'. $data['title'] . '</h1>';
-    ob_start();
-       @include $tpl;
-    $form = ob_get_contents();
-    ob_get_clean();
-    return $title. $form;
-}
+echo "Stage1";
 
 try {
-   $action =  isset($_GET['action'])? $_GET['action']  : 'controller_login';
+   $action =  isset($_GET['action'])? $_GET['action']  : 'controllerPage';
    $action();
 } catch (Exception $e) {
     echo "Не задан ";  // уровень отладки кода
