+++ model/ModelIndex.php
@@ -1,0 +1,8 @@
+<?php
+
+function model() {
+    // 1. Получить данные от сервера БД, файла,
+    //http , сайта, API
+    $arr = ['title'=> 'О нас','name'=>"Alisa"];
+    return $arr;
+}
