--- session/sess_fmv5rjfk3ojgrd3c2t1h3gtobr
@@ -1,1 +1,0 @@
-login|s:5:"admin";passwd|i:111;arr_num|a:5:{i:0;i:1;i:1;i:2;i:2;i:3;i:3;i:4;i:4;i:3;}
