+++ template/form.tpl
@@ -1,4 +1,4 @@
-<form action="auth.php" method="POST">
+<form action="../auth.php" method="POST">
     <label for="name">Ваше имя: <?php echo $name?> </label>
     <input name="login" id="name" type="text">
