+++ template/upload_form.tpl
@@ -1,4 +1,4 @@
-<form enctype="multipart/form-data" action="upload.php" method="post">
+<form enctype="multipart/form-data" action="../upload.php" method="post">
     <input type="hidden" name="MAX_FILE_SIZE" value="3000000" />
     <input name="userfile" type="file" />
     <input type="submit" value="Send File" />
